[move] 응답 통일 코드 경로 이동 #3
