+++ src/main/java/chungbazi/chungbazi_be/apiPayload/code/ErrorReasonDTO.java
@@ -1,4 +1,4 @@
-package chungbazi.chungbazi_be.apiPayload;
+package chungbazi.chungbazi_be.apiPayload.code;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
-public class ReasonDTO {
+public class ErrorReasonDTO {
 
     private HttpStatus httpStatus;
 
@@ -14,5 +14,5 @@ public class ReasonDTO {
     private final String code;
     private final String message;
 
-    public boolean getIsSuccess() {return isSuccess;}
+    public boolean getIsSuceess() {return isSuccess;}
 }
