+++ src/main/java/chungbazi/chungbazi_be/apiPayload/code/ReasonDTO.java
@@ -1,4 +1,4 @@
-package chungbazi.chungbazi_be.apiPayload;
+package chungbazi.chungbazi_be.apiPayload.code;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
-public class ErrorReasonDTO {
+public class ReasonDTO {
 
     private HttpStatus httpStatus;
 
@@ -14,5 +14,5 @@ public class ErrorReasonDTO {
     private final String code;
     private final String message;
 
-    public boolean getIsSuceess() {return isSuccess;}
+    public boolean getIsSuccess() {return isSuccess;}
 }
